compass-agents/requirements.txt
openai>=1.12.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
chromadb>=0.4.22
python-dotenv>=1.0.0
httpx>=0.27.0
